Non working version of DynamicContent alignment over multiple pages.

+++ OpenTemplater.Elements/IPositionedElement.cs
@@ -7,6 +7,6 @@
     {
         float XPosition { get; }
         float YPosition { get; }
-        byte ZOrder { get; }
+        int ZOrder { get; }
     }
 }
