+++ OpenTemplater.Elements/ImageElement.cs
@@ -5,7 +5,7 @@
         public string Key { get; set; }
         public float XPosition { get; set; }
         public float YPosition { get; set; }
-        public byte ZOrder { get; private set; }
+        public int ZOrder { get; private set; }
         public float Width { get; set; }
         public float Height { get; set; }
     }
