+++ OpenTemplater.Elements/LineElement.cs
@@ -6,7 +6,7 @@
         public string Key { get; private set; }
         public float XPosition { get; set; }
         public float YPosition { get; set; }
-        public byte ZOrder { get; private set; }
+        public int ZOrder { get; private set; }
         public float Width { get; set; }
         public float Height { get; set; }
         public IColor Color { get; set; }
